collector/requirements.txt
python-jobspy>=1.1.0
supabase>=2.0.0
openai>=1.0.0
python-dotenv>=1.0.0
requests>=2.31.0
pandas>=2.0.0
